♻️ Refactor: ui수정 및 className 삭제
